Add a manager interface to make it easier to write custom managers
We need to add this because we need a manager that don't depend on
monolog
